Simple handling of collisions with vertical walls
Adds simple handling of collisions with vertical walls.
This was done in order to give the possibility of further
development of character control mechanics

Changes the coordinates of the player object. Previously,
they pointed to the upper left corner of the hitbox,
now it is its center. The center of the hitbox is visible in-game

Refactors the player class constructor